refactor(BA-6790): move DB engine utilities from models to repository layer
Move engine creation/lifecycle functions (create_async_engine,
connect_database, vacuum_db) out of models/utils.py into a new
repositories/db/engine.py. This removes the models layer's dependency on
BootstrapConfig/DatabaseConfig and relocates engine ownership to the
repository layer.

The ExtendedAsyncSAEngine class and config-free transaction retry helpers
stay in models/utils.py since models row classes use them directly and
must not import from the higher repository layer.
