+++ tests/unit/manager/test_query_userinfo.py
@@ -47,8 +47,9 @@ from ai.backend.manager.models.runtime_variant import RuntimeVariantRow
 from ai.backend.manager.models.scaling_group import ScalingGroupRow
 from ai.backend.manager.models.session import SessionRow
 from ai.backend.manager.models.user import UserRole, UserRow
-from ai.backend.manager.models.utils import ExtendedAsyncSAEngine, create_async_engine
+from ai.backend.manager.models.utils import ExtendedAsyncSAEngine
 from ai.backend.manager.models.vfolder import VFolderRow
+from ai.backend.manager.repositories.db.engine import create_async_engine
 from ai.backend.manager.utils import query_userinfo, query_userinfo_from_session
 from ai.backend.testutils.db import with_tables
