+++ tests/unit/testutils/conftest.py
@@ -6,7 +6,8 @@ from collections.abc import Iterator
 import pytest
 
 from ai.backend.common.typed_validators import HostPortPair as HostPortPairModel
-from ai.backend.manager.models.utils import ExtendedAsyncSAEngine, create_async_engine
+from ai.backend.manager.models.utils import ExtendedAsyncSAEngine
+from ai.backend.manager.repositories.db.engine import create_async_engine
 from ai.backend.testutils.bootstrap import postgres_container  # noqa: F401
